+++ lib/contentful/management/client.rb
@@ -6,6 +6,7 @@ require 'contentful/management/version'
 require 'contentful/management/http_client'
 
 require 'contentful/management/client_space_methods_factory'
+require 'contentful/management/client_space_membership_methods_factory'
 require 'contentful/management/client_api_key_methods_factory'
 require 'contentful/management/client_asset_methods_factory'
 require 'contentful/management/client_content_type_methods_factory'
@@ -97,6 +98,15 @@ module Contentful
       # @return [Contentful::Management::ClientSpaceMethodsFactory]
       def spaces
         ClientSpaceMethodsFactory.new(self)
+      end
+
+      # Allows manipulation of space memberships in context of the current client
+      # Allows listing all space memberships for client and finding one by ID.
+      # @see _ README for details.
+      #
+      # @return [Contentful::Management::ClientSpaceMembershipMethodsFactory]
+      def space_memberships
+        ClientSpaceMembershipMethodsFactory.new(self)
       end
 
       # Allows manipulation of api keys in context of the current client