+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 * Added option to disable Content Type caching completely.
 * Added UI Extension Endpoint.
+* Added Space Memberships Endpoint.
 
 ## 1.8.1
 ### Added